Add a Dockerfile to the example application
To make it simpler for people to get up and running with the example I've added a Dockerfile so they can build a self contained docker image of the example application.

I've seen some questions on stack overflow from other people struggling to run the example, either because they don't know what port it runs on or because they forget to make it listen on 0.0.0.0 within a docker container. Both these issues bit me when I was trying to get started so I thought I'd add my Dockerfile to the project to help others.