Pass URI to socket.io-client to specify namespace
The example `app-namespace.py` which uses `templates/index.html`
did not correctly initialize the socket.io-client with the `/test`
namespace.  This prevented the client from communicating with
the example application.

This modification, constructs the URI based upon the current
`window.location`, and apppends `/test` so the client subscribes
to the correct namespace.